Use Feedback Helpers

This function allows a package developer to automatically add the suite of feedback msg_ functions utilized in this package to their own package.

Usage

use_feedback_helpers(pkg = ".", open = TRUE)

Arguments

pkg

Package - defaults to "."

open

boolean - should the newly created file be opened on creation?

Value

invisibly returns nothing - used for side-effects.

Details

Note that this adds a new file to your package under R/utils-msg_helpers.R as well as runs devtools::document() to document the added functions and their corresponding dependencies.

The functions added are:

msg_field()

msg_value()

msg_path()

msg_done()

msg_bullet()

msg_code()

msg_feedback()

msg_err()

msg_info()

plus, inform() and indent().

See also